@@ -14,7 +14,7 @@ description: A learning platform application for both web and mobile
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.6.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -29,6 +29,8 @@ dependencies:
   firebase_auth: ^0.8.3
   firebase_messaging: ^4.0.0+1
   #  google_sign_in: ^4.0.1+1
+  #  Carousel
+  carousel_slider: ^1.3.0
   geolocator: 5.0.1
   uuid: ^2.0.0
   image: ^2.0.7
@@ -38,6 +40,7 @@ dependencies:
   cached_network_image:
   flutter_svg:
   getflutter: ^1.0.11
+  universal_html: ^1.1.4
 
 dev_dependencies:
   flutter_test:
@@ -68,3 +71,5 @@ flutter:
     - assets/images/trainingpage/
     - assets/images/trainingpage/course-logos/
     - assets/images/upcomingpage/
+    - assets/images/aboutpage/
+    - assets/images/contactus/
