@@ -1,7 +1,9 @@
 import 'package:flutter/material.dart';
+import 'package:zengage_learning_platform/constants/app_colors.dart';
+import 'package:zengage_learning_platform/screens/assessment/assessment_page.dart';
 
 class BodyDropdownItemsList extends StatelessWidget {
-  final List<String> assessmentListItems;
+  final List<WhyItems> assessmentListItems;
   final String headerText;
 
   BodyDropdownItemsList(
@@ -28,7 +30,9 @@ class BodyDropdownItemsList extends StatelessWidget {
                 shrinkWrap: true,
                 itemCount: assessmentListItems.length,
                 itemBuilder: (context, index) {
-                  return AssessmentListItem(title: assessmentListItems[index]);
+                  return AssessmentListItem(
+                      title: assessmentListItems[index].title,
+                      content: assessmentListItems[index].contents);
                 }),
           ),
         ],
@@ -37,54 +41,40 @@ class BodyDropdownItemsList extends StatelessWidget {
   }
 }
 
-class AssessmentListItem extends StatefulWidget {
+class AssessmentListItem extends StatelessWidget {
   final String title;
+  final String content;
 
-  const AssessmentListItem({Key key, @required this.title}) : super(key: key);
-
-  @override
-  _AssessmentListItemState createState() =>
-      _AssessmentListItemState(title: title);
-}
-
-class _AssessmentListItemState extends State<AssessmentListItem> {
-  String title;
-
-  _AssessmentListItemState({this.title});
+  const AssessmentListItem({Key key, @required this.title, this.content})
+      : super(key: key);
 
   @override
   Widget build(BuildContext context) {
     return Padding(
       padding: const EdgeInsets.only(top: 40.0, bottom: 40.0),
-      child: GestureDetector(
-        onTap: () {
-          setState(() {
-            title = "Changed";
-          });
-        },
-        child: Container(
-          color: Colors.grey[300],
-          padding: EdgeInsets.fromLTRB(16.0, 26.0, 0.0, 26.0),
-          child: Row(
-            mainAxisAlignment: MainAxisAlignment.spaceBetween,
-            children: <Widget>[
-              Text(
-                title,
-                style: TextStyle(
-                    color: Colors.black45, fontWeight: FontWeight.bold),
-              ),
-              Padding(
-                padding: const EdgeInsets.only(right: 8.0),
-                child: Icon(
-                  Icons.add,
-                  semanticLabel: 'Expand',
-                  color: Colors.lightBlue,
-                  size: 44.0,
-                ),
-              )
-            ],
+      child: Column(
+        children: <Widget>[
+          Container(
+            width: MediaQuery.of(context).size.width * 0.80,
+            color: Colors.grey[300],
+            padding: EdgeInsets.fromLTRB(16.0, 26.0, 0.0, 26.0),
+            child: Text(
+              title,
+              style:
+                  TextStyle(color: textGreyColor, fontWeight: FontWeight.bold),
+            ),
           ),
-        ),
+          Container(
+            color: Colors.white,
+            padding: EdgeInsets.fromLTRB(16.0, 26.0, 0.0, 26.0),
+            child: Text(
+              content,
+              style: TextStyle(
+                color: Colors.black, /*fontWeight: FontWeight.bold*/
+              ),
+            ),
+          )
+        ],
       ),
     );
   }
