@@ -1,7 +1,9 @@
 import 'package:flutter/material.dart';
 
-final Color textGreyColor = Color(0xff6B6B7E);
+const Color textGreyColor = Color(0xff6B6B7E);
 
-final Color bgGreyColor = Color(0xffEDEDF5);
+const Color bgGreyColor = Color(0xffEDEDF5);
 
-final Color darkGreyColor = Color(0xff6B6B7E);
+const Color darkGreyColor = Color(0xff6B6B7E);
+
+const Color blueThemeColor = Color(0xff2A96D9);
