@@ -1,4 +1,4 @@
-final String appName = 'Zengage';
+final String appName = 'Agile TechPrime';
 
 final String homePageTitle = 'Welcome';
 final String routeErrorTitle = 'Error';
@@ -6,7 +6,7 @@ final String routeErrorTitle = 'Error';
 final String signUp = 'Sign Up';
 final String signIn = 'Sign In';
 
-final String copyrightText = '@Tech Prime 2020 All Rights Reserved';
+final String copyrightText = '@$appName 2020 All Rights Reserved';
 
 final String coachingBodyText =
     'Want to introduce the power of Agile to your organization? '
