@@ -17,9 +17,18 @@ class _CoachingPageState extends State<CoachingPage> {
   @override
   Widget build(BuildContext context) {
     final coachingList = [
-      'Agile Transformation',
-      'DevOps Transformation',
-      'Tools Assessments & Implementation'
+      WhyItems(
+          title: 'Agile Transformation',
+          contents:
+              'Want to introduce the power of Agile to your organization?  Our coaches can work with your teams to bring them up to speed.  We will define the roles and responsibilities of the various team members and stakeholders, assist with implementing Agile practices including Scrum, Kanban and Lean, and help to position your group into moving ahead on your own.'),
+      WhyItems(
+          title: 'DevOps Transformation',
+          contents:
+              'Combining both software development and operations, an Agile DevOps environment maximizes productivity by coordinating a shared vision.  This enables your organization to cooperate on a common set of tools and practices, from planning to maintenance.'),
+      WhyItems(
+          title: 'Tools Assessment and Implementation',
+          contents:
+              'Agile is a set of principles – there are a plethora of tools available, but which ones are best for your specific needs and goals?  With this coaching session, we help you assess your current toolset and your organization’s strengths and challenges, and we assemble a collection of tools specific to your requirements.')
     ];
 
     return SafeArea(
@@ -32,7 +41,6 @@ class _CoachingPageState extends State<CoachingPage> {
                 children: <Widget>[
                   ContactUsImageBanner(
                     imageLink: 'assets/images/coaching/contact-us-coaching.jpg',
-                    isButtonTranslucent: true,
                   ),
                   WhyBodyHeader(
                     topImageLink:
