@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:zengage_learning_platform/constants/app_colors.dart';
 
 ///Contains all the information related to styling of the app
 ///
@@ -14,7 +15,7 @@ const AppBarStyle = TextStyle(
     fontWeight: FontWeight.w300,
     fontSize: MediumTextSize,
     backgroundColor: Colors.white,
-    color: Colors.blue);
+    color: blueThemeColor);
 
 const TextTitleStyle = TextStyle(
     fontFamily: FontNameDefault,
