@@ -1,5 +1,7 @@
 import 'package:flutter/material.dart';
+import 'package:zengage_learning_platform/extensions/hover_extensions.dart';
 import 'package:zengage_learning_platform/routes/route_generator.dart';
+import 'package:zengage_learning_platform/widgets/body/BlueBorderWhiteButton.dart';
 
 class MissionContainer extends StatelessWidget {
   final String imagePath;
@@ -15,49 +17,69 @@ class MissionContainer extends StatelessWidget {
       decoration: BoxDecoration(
           image:
               DecorationImage(image: AssetImage(imagePath), fit: BoxFit.cover)),
-      child: Row(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        mainAxisAlignment: MainAxisAlignment.spaceEvenly,
+      child: Column(
+        mainAxisAlignment: MainAxisAlignment.spaceBetween,
         children: <Widget>[
-          Column(
-            crossAxisAlignment: CrossAxisAlignment.start,
-            mainAxisAlignment: MainAxisAlignment.center,
-            children: <Widget>[
-              MissionButton(
-                buttonText: "Who We are",
-              ),
-              MissionButton(
-                buttonText: "What We do",
-              ),
-            ],
-          ),
           Container(
-            child: Column(
+            margin: EdgeInsets.only(
+                top: MediaQuery.of(context).size.height / 2 * 0.2),
+            child: Row(
               crossAxisAlignment: CrossAxisAlignment.start,
-              mainAxisAlignment: MainAxisAlignment.center,
+              mainAxisAlignment: MainAxisAlignment.spaceEvenly,
               children: <Widget>[
-                Text(
-                  "Our Mission:",
-                  style: TextStyle(
-                      color: Colors.black,
-                      fontSize: 24.0,
-                      fontWeight: FontWeight.w800),
+                Column(
+                  crossAxisAlignment: CrossAxisAlignment.start,
+                  mainAxisAlignment: MainAxisAlignment.center,
+                  children: <Widget>[
+                    MissionButton(
+                      buttonText: "Who We are",
+                    ),
+                    MissionButton(
+                      buttonText: "What We do",
+                    ),
+                  ],
                 ),
                 Container(
-                  width: MediaQuery.of(context).size.width / 4,
-                  height: MediaQuery.of(context).size.height / 4,
-                  child: Text(
-                    "We will guide you and your organzation in your quest to implement an Agile environment. We will help you assess your current state, provide training to elevate your knowledge of Agile tools and practices, and offer individualized coaching as you roll out Agile. We are dedicated to helping you succeed!",
-                    style: TextStyle(
-                        color: Colors.black,
-                        fontSize: 20.0,
-                        fontWeight: FontWeight.w500),
+                  child: Column(
+                    crossAxisAlignment: CrossAxisAlignment.start,
+                    mainAxisAlignment: MainAxisAlignment.center,
+                    children: <Widget>[
+                      Text(
+                        "Our Mission:",
+                        style: TextStyle(
+                            color: Colors.black,
+                            fontSize: 24.0,
+                            fontWeight: FontWeight.w800),
+                      ),
+                      Container(
+                        width: MediaQuery.of(context).size.width / 4,
+                        height: MediaQuery.of(context).size.height / 4,
+                        child: Text(
+                          "We will guide you and your organzation in your quest to implement an Agile environment. We will help you assess your current state, provide training to elevate your knowledge of Agile tools and practices, and offer individualized coaching as you roll out Agile. We are dedicated to helping you succeed!",
+                          style: TextStyle(
+                              color: Colors.black,
+                              fontSize: 20.0,
+                              fontWeight: FontWeight.w500),
+                        ),
+                      ),
+                    ],
                   ),
-                ),
-                buildCollabButton(context, RouteGenerator.TRAINING_ROUTE),
+                )
               ],
             ),
-          )
+          ),
+          Align(
+            alignment: Alignment.bottomRight,
+            child: Container(
+              margin: EdgeInsets.only(
+                  bottom: MediaQuery.of(context).size.height / 2 * 0.1,
+                  right: MediaQuery.of(context).size.height / 2 * 0.1),
+              child: BlueBorderWhiteButton(
+                buttonText: "COLLABORATE",
+                route: RouteGenerator.TRAINING_ROUTE,
+              ),
+            ),
+          ),
         ],
       ),
     );
@@ -95,25 +117,27 @@ class MissionButton extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return Container(
-      width: MediaQuery.of(context).size.width / 4,
-      child: Padding(
-        padding: EdgeInsets.symmetric(vertical: 16.0),
-        child: Material(
-          color: Colors.lightBlue,
-          elevation: 5.0,
-          child: MaterialButton(
-            onPressed: () {
-              // Implement login functionality.
-            },
-            minWidth: 200.0,
-            height: 72.0,
-            child: Text(
-              buttonText,
-              style: TextStyle(
-                  color: Colors.white,
-                  fontSize: 32.0,
-                  fontWeight: FontWeight.bold),
+    return HandCursor(
+      child: Container(
+        width: MediaQuery.of(context).size.width / 4,
+        child: Padding(
+          padding: EdgeInsets.symmetric(vertical: 16.0),
+          child: Material(
+            color: Colors.lightBlue,
+            elevation: 5.0,
+            child: MaterialButton(
+              onPressed: () {
+                // Implement login functionality.
+              },
+              minWidth: 200.0,
+              height: 72.0,
+              child: Text(
+                buttonText,
+                style: TextStyle(
+                    color: Colors.white,
+                    fontSize: 32.0,
+                    fontWeight: FontWeight.bold),
+              ),
             ),
           ),
         ),
