@@ -1,13 +1,16 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
+import 'package:zengage_learning_platform/widgets/body/BlueBorderWhiteButton.dart';
 
 class SliderContent extends StatelessWidget {
   final String imagePath;
   final String headlineText;
   final String buttonText;
   final String route;
+
   SliderContent(
       {this.imagePath, this.headlineText, this.buttonText, this.route});
+
   @override
   Widget build(BuildContext context) {
     return Container(
@@ -17,9 +20,11 @@ class SliderContent extends StatelessWidget {
         image: DecorationImage(image: AssetImage(imagePath), fit: BoxFit.cover),
       ),
       child: Padding(
-        padding: EdgeInsets.only(top: 75.0, right: 50.0),
+        padding: EdgeInsets.only(top: 75.0, bottom: 45.0, right: 50.0),
         child: Container(
           child: Column(
+            crossAxisAlignment: CrossAxisAlignment.end,
+            mainAxisAlignment: MainAxisAlignment.spaceBetween,
             children: <Widget>[
               Text(
                 headlineText,
@@ -31,26 +36,10 @@ class SliderContent extends StatelessWidget {
               SizedBox(
                 height: 50.0,
               ),
-              Container(
-                width: MediaQuery.of(context).size.width / 7,
-                height: MediaQuery.of(context).size.height / 13,
-                child: FlatButton(
-                  shape: RoundedRectangleBorder(
-                      borderRadius: BorderRadius.circular(1.0),
-                      side: BorderSide(color: Colors.blue)),
-                  color: Color(0xf6f8fa),
-                  textColor: Colors.white,
-                  padding: EdgeInsets.all(8.0),
-                  onPressed: () {
-                    Navigator.pushNamed(context, route);
-                  },
-                  child: Text(
-                    buttonText.toUpperCase(),
-                    style: TextStyle(
-                      fontSize: 22.0,
-                    ),
-                  ),
-                ),
+              Align(
+                alignment: Alignment.bottomRight,
+                child: BlueBorderWhiteButton(
+                    buttonText: buttonText.toUpperCase(), route: route),
               ),
             ],
           ),
